+++ bin/slaw
@@ -23,6 +23,7 @@ class SlawCLI < Thor
   option :section_number_position, enum: ['before-title', 'after-title', 'guess'], desc: "Where do section titles come in relation to the section number? Default: before-title"
   option :reformat, type: :boolean, desc: "Reformat common formatting issues to make grammar matching better. Default: true for PDF files, false otherwise"
   option :crop, type: :string, desc: "Crop box for PDF files, as 'left,top,width,height'."
+  option :grammar, type: :string, desc: "Grammar name (usually a two-letter country code). Default is za."
   def parse(name)
     logging
 
@@ -50,7 +51,7 @@ class SlawCLI < Thor
       text = extractor.extract_from_file(name)
     end
 
-    generator = Slaw::ActGenerator.new
+    generator = Slaw::ActGenerator.new(options[:grammar] || 'za')
 
     text = generator.reformat(text) if options[:reformat]
